+++ o/soft_robot/derived/ikko_dake/eqs/c_src/A/A_7_1.h
@@ -10,7 +10,7 @@
 #ifndef A_7_1PROJECT__A_7_1__H
 #define A_7_1PROJECT__A_7_1__H
 
-int A_7_1();
+double A_7_1();
 
 #endif
